requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
accelerate 
pypdf
python-dotenv
openai
langdetect
dotenv
bs4
pdfplumber
langchain
langchain-community
langchain-huggingface
langchain_ollama
langchain_openai
Scikit-learn
streamlit_mic_recorder 
streamlit_option_menu
fpdf
PyPDF2
python-docx
docx2txt
sqlalchemy


